+++ components/_app/listing/main-info-col.tsx
