+++ components/_app/listing/main-info.tsx
